python/requirements.txt
# Python Dependencies for Geospatial Data Management
# Core geospatial libraries
geopandas==0.14.1
shapely==2.0.2
folium==0.15.1
geopy==2.4.1
pyproj==3.6.1

# Data processing and analysis
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
scikit-learn==1.3.2

# Machine learning for spatial analysis
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.3.4

# Web framework for Python API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Database connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Data import/export
fiona==1.9.5
rasterio==1.3.9
geopandas==0.14.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
